Moving finder tests into a single file
