+++ ingestion_tools/scripts/tests/s3_import/test_finders.py
@@ -1,11 +1,15 @@
+import json
+
 import pytest
 from importers.dataset import DatasetImporter
 from importers.run import RunImporter
+from importers.tiltseries import TiltSeriesImporter
 from importers.utils import IMPORTERS
 from importers.voxel_spacing import VoxelSpacingImporter
+from mypy_boto3_s3 import S3Client
 
 from common.config import DepositionImportConfig
-from common.finders import DefaultImporterFactory
+from common.finders import DefaultImporterFactory, DestinationFilteredMetadataFinder
 from common.fs import FileSystemApi
 
 
@@ -132,3 +136,47 @@ def test_multi_parent_filters(deposition_config, parents, source_config):
     finder = DefaultImporterFactory(source_config, VoxelSpacingImporter)
     items = finder.find(deposition_config, {}, **parents)
     assert len(items) == 1
+
+
+@pytest.mark.parametrize(
+    "filters, expected_prefixes",
+    [
+        ([{"key": "non-existent-field-name", "value": 300}], []),
+        ([{"key": "deposition_id", "value": 300}], [100]),
+        ([{"key": ["authors", "name", "first"], "value": "Bob"}], [101]),
+        ([{"key": ["authors", "name", "first"], "value": "Alice"}], []),
+        ([{"key": ["invalid", "name", "first"], "value": "Bob"}], []),
+        ([{"key": "deposition_id", "value": 303}], []),
+        ([], [100, 101]),
+    ],
+)
+def test_destination_filtered_metadata_finder(
+    s3_fs: FileSystemApi,
+    s3_client: S3Client,
+    test_output_bucket: str,
+    filters: list[dict],
+    expected_prefixes: list[int],
+):
+    config_file = "tests/fixtures/dataset1.yaml"
+    config = DepositionImportConfig(s3_fs, config_file, test_output_bucket, "test-public-bucket", IMPORTERS)
+    dataset_id = 20000
+    run_name = "random_run"
+
+    def _put_data(prefix: int, data: dict):
+        key = f"{dataset_id}/{run_name}/TiltSeries/{prefix}-tiltseries_metadata.json"
+        s3_client.put_object(Bucket=test_output_bucket, Key=key, Body=json.dumps(data).encode("utf-8"))
+
+    _put_data(100, {"deposition_id": 300})
+    _put_data(
+        101,
+        {"deposition_id": 301, "authors": [{"name": {"initial": "A", "last": "Smith"}}, {"name": {"first": "Bob"}}]},
+    )
+
+    finder = DestinationFilteredMetadataFinder(filters, importer_cls=TiltSeriesImporter)
+    actual_result = finder.find(config, {"dataset_name": dataset_id, "run_name": run_name})
+
+    expected = {}
+    for expected_prefix in expected_prefixes:
+        filename = f"{test_output_bucket}/{dataset_id}/{run_name}/TiltSeries/{expected_prefix}-tiltseries_metadata.json"
+        expected[filename] = filename
+    assert actual_result == expected
